Check if the interface link is connected on network test failure.
If a network test which applies customized networking with an interface
parameter fails maas-run-remote-scripts will check if the link is still
connected. If it is not the result YAML file will be appended with
link_connected: False. When the result YAML is processed MAAS will
automatically change the interface's link_connected status.