Release AUTOIP addresses when powered off in a non-use status
Commissioning and testing now require AUTOIP addresses to be assigned to
allow for network validation testing. The signal to end these statuses
comes before the machine is actually turned off. Further users may keep
the machine on after commissioning/testing to debug script issues. To
ensure that AUTOIPs are released MAAS will now release any assigned
AUTOIP if the machine isn't currently in a status that needs an IP and
is off.