Fixes LP: #1841807 - Handle the case where the RPC connection can be lost in a list-boot-images API call.
There is a case where the RPC connection can be lost for multiple reasons during an RPC call to a rack controller. The error path was not handling this case, but was handling other cases. This updates that error path to handle the connection lost case so the API output is consistent and not an improper error or API return code.