LP: #1855035 - Run network tests against the boot interface after commissioning
When network configuration is applied to the boot interface no configured IP
address is associated with the device. The ParametersForm only allows
configured devices to be tested so the boot interface is ignored causing the
place holder to be left. During commissioning the ParametersForm can safely
allow the boot interface to be tested as we know it works with DHCP.