LP: #1855022 - Cleanup orphaned ScriptResults after commissioning failure.
Scripts which have parameters may become orphaned when created before
commissioning has run when an error occurs during commissioning.
ScriptResults on an uncommissioned machine use the 'all' place holder which is mapped to devices after commissioning has run. If commissioning
fails this mapping never occurs. When a subsequent commissioning passes
the mapping works and the unmapped ScriptResult are never cleaned up as
they appear unique. This leaves failed tests associated with the node.