LP #1794882: dhcpd doesn't start until maas-rackd is restarted
The code that registers the listener for dhcp events were called from
multiple thread, and wasn't threadsafe.

Considering that we shouldn't do db work inside the main twisted loop, I
refactored the code so that register a listener only modifies the
listener dicts. Then a separate looping call goes through the list of
channels to listen to and issue the db call in a separate thread.

That removed the race condition where the rack controller service would
register itself to listen for dhcp events, but the listen wouldn't
actually listen to that channel.

This branch also makes use of the fact that we now use the asyncio
reactor and can thus use async function definitions.