LP #1873430:  Subnet pages take a long time to load.
Speed up space.list websocket handler when there are many ip addresses.

The space.list handler prefetched staticipaddress, which make scales
very badly with the number of statipaddress records.

Added tests to ensure that the number of queries is constant and
modified the prefetch and logic to make it pass.

Also remove unneeded prefetch for fabric.list websocket call.