Require the mode to be passed as a positional argument.
The idea is that when you run 'maas init' without any arguments, it will
explain how to initialise MAAS. Each mode/subcommand (region/rack) has
only the argument it needs, so that it's easier to know what to pass
when adding a rack server, for example.

The old --mode way of calling 'maas init' is still supported (including
passing 'all'), but will be deprecated for 2.8 in another branch.