Fix 'maas config' so that it doesn't break when showing the config.
I also removed code for changing the mode, since it's too hard to
maintain backwards-compatibility. If the mode needs to be changed, 'maas
init' should be used.

I also didn't bother supporting the --database-uri parameter, since
'maas config' is currently quite low-level and just writes the passed
parameters to the config directly. In the future we should probably get
rid if 'maas config' completely and replace it with 'snap config'.