Add a new power-driver subcommand to maas-rack
maas-rack power-driver is a new utility CLI command for directly controlling the power driver and using them to power on, off or
cycle a system.

As a example usage, the following would power on a virsh machine.

$ maas-power on virsh --power-address "qemu+ssh://root@127.0.0.1/system" --power-id expert-grub

Drivers have different arguments requirements which can be seen by looking at the help output. For example;

$ maas-power on ipmi --help