Remove the virsh and lxd power drivers and error on registry overwrites
Previously if a key was reused in the registry then the existing item
would be silently replaced with the new one. Now unregister_item needs
to be used before replacing a key.

The old behaviour led to the virsh and lxd power drivers being
registered but unused (as they were replaced by pod drivers with the
same name). They have been removed.