Fix up the broken pod API tests.
forms.PodForm.discover_and_sync_pod adds a task to the reactor via
callLater() to do things post commit.

The tests don't have a transaction boundary in the same way as the
form code expects, so we mock out the post_commit_do. The behaviour of
the form can be tested by the form tests.

Drive-by fixups to separate tests by if they're needing a regular user
or an admin, and avoid running the tests twice to test different
authentication mechanisms.