Fix tests that fail intermittently.
The test_storage_single_contraint_allows_root_on_virtual failed because the
partition size could be random and as low as MIN_BLOCK_DEVICE_SIZE, which is
not enough if you want to create a VG and a block device inside it.

The other tests failed due to test isolation issues.