Allow ovs bridges to be used with latest netplan.
The netplan support is currently in preview and not backported to focal
yet.

The MAAS support is also considered to be preview only, and we'll need to clean it up when netplan is backported to focal and to clean up the experience.

Currently, no parameters are supported for ovs bridges.