Make LXD discover method use asyncio.
This exposed the fact that getting the list of networks was done in the main
thread, which now has been fixed.

With @inlineCallbacks it's hard to ensure that you don't do blocking calls in
the main thread, since 'yield' doesn't complain if it doesn't get a deferred.

The code in question was untested, so I amended the existing test to exercise
that part of the code. I'll look into improving the testing story in another
branch.