Owner Data / Workload Annotations
Add support for the Ownerdata model in the machine websocket handler.
This will let the UI set and get owner data for a machine.

Setting owner data is a simple key value mapping, and getting it will
return that value for the machine. Keys can be removed by setting them
with a empty string.