Stop a partition before starting a paused or terminated partition.
When an IBM Z partition goes into a paused or terminated state it must be
stopped before it can be started. MAAS has to hold a thread on the rack
controller when this is detected while waiting for the partition to stop
before it can issue the start call. This isn't ideal and IBM is aware we
would like it changed.