Refresh the commission output for the region in the network monitoring service.
Now both the rackd and regiond uses the network interfaces monitoring
service to update the commissioning output.

This change is needed, so that I can include the hints in the
commissioning output, so that we can update all the interface details in
the commissioning hooks, and get rid of Controller.update_interfaces().