Move the logic for Controller.update_interface() to its own file.
We want to use the same logic for non-controller nodes, and eventually merge it with similar logic in metaserver.builtin_scripts.hooks.py.

This commit simply moves the methods to their own file, so that it's easier to
see how much it actually is.

This is pretty much a straight move. The only thing I did was to remove _ from
the function names and replace 'self' with 'node', as well as changing some
mocking setup in the tests.