Send the beaconing hints in the commissioning data.
I forgot to include them when I initially added the information that the
networks interfaces monitoring service collects to the commissioning output.

As before, don't pay too much attention on the key names. The will most likely
change soon, as I make use of the data, and remove what's not needed.