Format annotated commissioning data better.
When the network interface monitoring service sends commissioning
output, it annotates it with more information. This commit makes it so
that the commissioning data stays nicely formatted, rather than having
it be a single long line.

It also updates the combined output, since that's the default view when
looking at the commissioning data.