Prepare for 3.0.0b2 release and update utilites/release-prepare to new process
Since we tag after the release, it doesn't make sense of release-prepare
to tag the commit at this point.

The python version now resides in setup.cfg, so the regexp was updated
to cope with it.

I also changed it so that DEBFULLNAME and DEBEMAIL are automatically set
(if not set already) to the name and email in the git config, which is
most likely what you want.

In addition to this, the commit is now show, so that you can verify it
did the right thing.