Convert most update_node_interfaces tests to use FakeCommissioningData.
There's a small change to FakeCommissioningData itself, to fix a few
bugs and add simple support for storage, to make the tests pass.

The remaining tests are for hints and monitored interface, which I will
convert in another branch.

Since I was touching the asserts anyway, I took the opportunity to
remove testtools from them as well.

I removed couple of tests that didn't have any purpose, like testing the
ordering of calls, as well as a test that didn't have any asserts at all.