Allow update_node_interfaces() to process non-controller machines.
For now I made it so that deployed nodes work the same as controllers,
so that all new interfaces are marked as 'acquired'.

Then for non-deployed machines, I assume that they are commissioning and
new physical interfaces are not marked as acquired.

I kept all other types of interfaces as 'acquired' while commissioning,
but we might want allow users to set up the networking in a
commissioning script, so that things like vlan interfaces and bonds
might be created automatically.