LP #1926140: maas_url not returned to the UI
The UI gets the URL from the config websocket handler, but the code in
there to return the maas url was broken. The maas url was never stored
in the database, so the UI always got the default
http://localhost:5240/MAAS.

Now we sync the maas_url from regiond.conf to the database on startup.
Eventually we want to get rid of maas_url from regiond.conf completely
and store it only in the database.