Add tags and 'default' information to maas-scripts metadata endpoint.
Also added "deploy-info" tag to the builtin scripts that are marked as being
run on a controller. The idea is that when someone gets the scripts from the
metadata server, they can choose to run only the same scripts that MAAS itself
runs.

Where possible, I updated the callsites to use the information in the database,
rather than looking at NODE_INFO_SCRIPTS. runscripts() couldn't be updated,
though, since it doesn't have access to the database. In the future, it should
be modified to download the scripts from the region, so that it gets the
metadata that way.