Add a DB patch to migrate old controllers to DEPLOYED.
In 3.0, and previous versions, whenever a node was created to track the
state of the controller, it was created in a NEW state.

Now we created such nodes as DEPLOYED, and this patch migrates to
controllers that were create in 3.0 and below.

If a controller is linked to a machine that wasn't automatically
created, it's not updated.

I also updated factory.make_RegionController to default to creating the
node as DEPLOYED. make_RegionRackController and  make_RackController
already defaulted to DEPLOYED.