Fix dblocks intermittent test failure.
Postgres advisory locks are global across sessions and databases, which
meant that other tests that run in parallel might acquire and release
locks.

The test_lock_actually_locked tests did grab an exclusive lock for the
cluster, but most tests using locks didn't respect that lock.

I changed it so that the test only looks at the lock it's interested in,
and don't care if there are additional locks.