Don't run complex and expensive post commit hooks during testing.
The delete_large_object_content_later and cache_boot_sources were being
called in post commit hooks in signals. In general, it's hard to know
when they are being called, since the signals are kind of hidden away.

This disables the post commit hooks in the bootsources and largefile
signals, unless the test explicitly asks for them to be enabled.

We've had intermittent test failurs around boot sources, where the
transaction gets in a broken state, as well as some mocks not being
called.

I've run the test suite quite many times now with this fix, and I
haven't seen it fail so far.