Fix filedescriptor bug in secure_erase_hdparam()
Close the filedescriptor through its file object, instead of using
os.close.

The relevant code created a file descriptor using os.open, and then
creates a file object for the descriptor using os.fdopen. It then closed
the file descriptor using os.close(), which meant that the file object
didn't know that the file descriptor had been closed.

Then when the garbage collector delete the file object, it tried to
close the file descriptor again, and got a "bad file descriptor" error.

Not sure if this affected a running MAAS, but we say quite frequent
intermittent test failures due to this.