fix RAID device available space reporting
MAAS was reporting a misleading value for the available space in RAID virtual devices, caused by a missing Partition table. Creating the table was a side effect of creating the first partition, and at this point the UI already tried to use the full size of the device.

fixes LP#1938296