LP:1972052  fix VirtualMachineInterfaceCurrentConfigManager query filter
The Pod.hints.hosts is not always populated for VM hosts (e.g. for virsh hosts), but the VirtualMachineInterface can still be linked to a host interface in that case.
This changes the query filter to go through the host interface instead of the hints.

It also includes a change to log the exception cause by a Conflict HTTP
response, which was hiding the original issue in the linked bug.