Remove unused code from Node.set_power_config().
There was code that checked for self.bmc, even though it couldn't really
be None, except for in unit tests.

A lot of tests passed in "" as the power type, expecting to get a node
without a BMC configured. But in reality, that created a BMC with power
type "", which is invalid.