Explicity delete device children only when release a machine.
Previously, only devices could be a child of a machine. Now that
machines (like VMs and DPUs) can be children, we need to explicitly
filter out the non-device nodes.

DPUs shouldn't be deleted, and VMs will be deleted when the VM
host is deleted. There are checks that prevents a VM host from
being released.